==> 1 Data Wrangling and Subsetting.ipynb <==
# Import libraries
import pandas as pd
import numpy as np
import os

# Import orders.csv
path = r'/Users/elvachen/Documents/DA Learning/Immersion-data resources/original data'
df_order = pd.read_csv(os.path.join(path, 'orders.csv'),index_col = False)

# Get rid of column 'eval_set'
df_order.drop(columns = ['eval_set'])

# Update to dataframe without column 'eval_set' 
df_order = df_order.drop(columns = ['eval_set'])

# Check the reseult
df_order.head(20)

# List the column 'days_since_prior_order'
df_order['days_since_prior_order'].value_counts(dropna = False)

# Rename column 'order_dow'
df_order.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

# Check the result
df_order.head(20)

# Change data type of column 'order_id'
df_order['order_id'] = df_order['order_id'].astype('str')

# Check the reslut
df_order.dtypes

# Import 'departments.csv'
df_dept = pd.read_csv(os.path.join(path, 'departments.csv'),index_col = False)

# Check the result
df_dept.head(20)

# Transpose the df
df_dept=df_dept.T

df_dept

# Reset index
df_dept.reset_index()

# Set df_dept with new column and get rid of the index
new_header=df_dept.iloc[0]
df_dept=df_dept.iloc[1:]
df_dept.columns=new_header

# Check the result
df_dept

Task of 4.4

# 4.2.2 Change the data type of column 'user_id' in df_order
df_order['user_id'] = df_order['user_id'].astype('str') 

# 4.2.3 Rename column 'order_number'
# Rename column 'order_dow'
df_order.rename(columns = {'order_number' : 'order_sequency'}, inplace = True)

# 4.2.4 Find out busiest hour for placing order
df_order['order_hour_of_day'].value_counts(dropna = False) 

The busiest hour for placing order is 10 am.

# 4.2.5 Determine the meaning behind a value of 4 in the "department_id"
data_dict = df_dept.to_dict('index')
print(data_dict.get('4'))


# 4.2.6 Create a subset containing only breakfast item.
df_prod = pd.read_csv(os.path.join(path, 'products.csv'),index_col = False)
df_breakfast = df_prod.loc[df_prod['department_id'] == 14]

# Review the result
df_breakfast

# 4.4.7 Find all observations from the entire dataframe that include items from the following departments: 
#       alcohol, deli, beverages, and meat/seafood.
df_dinnerparty = df_prod.loc[df_prod['department_id'].isin([5, 20, 7, 12])] 

# Review the result
df_dinnerparty

# 4.4.8 How many rows does the last dataframe you created have?
# There are 7650 rows in dataframe 'df_dinnerparty'.

# 4.4.9 Extract all the info with "user_id" of "1"
df_user_id_1 = df_order.loc[df_order['user_id']=='1']

# Review the result
df_user_id_1

# 4.4.10 Basic stats of the user's behavior
df_user_id_1.describe()

# 4.4.12 Export df_ords dataframe as "orders_wrangled.csv" in your "prepared Data" folder
df_order.to_csv(os.path.join(path, 'Prepared Data', 'orders_wrangled.csv'))

# 4.4.13 Export df_dep_t_new dataframe as "departments_wrangled.csv"
df_dept.to_csv(os.path.join(path, 'Prepared Data', 'departments_wrangled.csv'))
