==> 2 Consistency Checks.ipynb <==
# Examples throughout the lesson

# Import libraries
import pandas as pd
import numpy as np
import os

# Create a dataframe
df_test=pd.DataFrame()

# Create a mixed type column
df_test['mix']=['a','b', 1, True]

# Review the result
df_test

# Check if the column contains different data types
for col in df_test.columns.tolist():
  weird = (df_test[[col]].applymap(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)

# Change the mixed-data-type column to string
df_test['mix'] = df_test['mix'].astype('str')

# Import productsl.csv
path = r'/Users/elvachen/Documents/DA Learning/Immersion-data resources'
df_prods = pd.read_csv(os.path.join(path, 'original data', 'products.csv'),index_col = False)

# Look for the null values
df_prods.isnull().sum()

# Create a dataframe to contain the null values
df_nan=df_prods[df_prods['product_name'].isnull()== True]
df_nan

# Check the shape of the dataframe
df_prods.shape

# Create a clean dataframe without the null-value rows
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

# Check the shape of the clean dataframe
df_prods_clean.shape

# Look for duplicates 
df_dups = df_prods_clean[df_prods_clean.duplicated()]

# Drop duplicates
df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

# Check the shape of the new dataframe
df_prods_clean_no_dups.shape

# Export the clean dataframe to csv file
df_prods_clean_no_dups.to_csv(os.path.join(path, 'prepared data', 'products_clean.csv'))

# Task of 4.5

# 4.5.2 Anything about the data looks off or should be investigated further.
df_prods.describe()

The maximum price is unusual which is 99999 when the 3/4 quartile is 11.2. 

# 4.2.3 Check for mixed-type data in your df_ords dataframe.
df_order_wrangled = pd.read_csv(os.path.join(path, 'prepared data', 'orders_wrangled.csv'),index_col = False)
for col in df_order_wrangled.columns.tolist():
    weird = (df_order_wrangled[[col]].applymap(type) != df_order_wrangled[[col]].iloc[0].apply(type)).any(axis = 1)
    if len (df_order_wrangled[weird]) > 0:
        print (col)

# 4.2.4 If you find mixed-type data, fix it.

There is no mixed-type data found.

# 4.2.5 Run a check for missing values in your df_ords dataframe.
df_order_wrangled.isnull().sum()

According to the result above, there is only one column "dats_since_prior_order" that has missing data. The number of missing values in the column is 206209. It makes sences customers who just placed only one order haven't no values for "days_since_prior_order". 

# 4.2.6 Address the missing values using an appropriate method.

To check if all customers who placed more than one order have respectively values in the column "days_since_prior_order", if not, fixed them with calculation of the durations between appropriate oreder dates. Then, leave the missing values in the column for all first order of the customers in the dataframe, as they interprets correct meaning of the business.

# 4.2.7 Run a check for duplicate values in your df_ords data.
df_order_wrangled[df_order_wrangled.duplicated()]

According to the above result, no duplicated rows found.

# 4.2.8 Address the duplicates using an appropriate method.

No duplicates were found in the dataframe.

# 4.2.9 Export your final, cleaned df_prods and df_ords data as ".csf" files in your "Prepared Data" folder.
df_order_wrangled.to_csv(os.path.join(path, 'prepared data', 'orders_clean.csv'))
