==> 3 Preparing Data - Combining Datasets.ipynb <==
## Merging Instacart Data

# Import libraries
import pandas as pd
import numpy as np
import os

# Import orders_clean.csv
path = r'/Users/elvachen/Documents/DA Learning/Immersion-data resources'
df_orders = pd.read_csv(os.path.join(path, 'prepared data', 'orders_clean.csv'),index_col = False)

# Import order_products_prior.csv
path = r'/Users/elvachen/Documents/DA Learning/Immersion-data resources'
df_orders_prior = pd.read_csv(os.path.join(path, 'original data', 'order_products_prior.csv'),index_col = False)

# Check the output
df_orders_prior.head()

# Check the shape of dataframe
df_orders.shape

# Check the shape of dataframe
df_orders_prior.shape

# Merge the dataframes with inner join
df_merged_large = df_orders.merge(df_orders_prior, on = 'order_id', indicator = True)
df_merged_large.head()

# Count the values of column "_merge"
df_merged_large['_merge'].value_counts()

# Merge the dataframes with outer join
df_merged_large = df_orders.merge(df_orders_prior, on = 'order_id', how='outer', indicator = True)

# Count the values of column "_merge"
df_merged_large['_merge'].value_counts()

## Exercise of 4.6

# 4.6.2 Export the merged file in pickle format as “orders_products_combined.pkl”
df_merged_large.to_pickle(os.path.join(path, 'Prepared Data', 'orders_products_combined.pkl'))
