==> 4 Analyzing - Deriving New Variables.ipynb <==
## Examples of the Class

# Import libraries
import pandas as pd
import numpy as np
import os

# Import "orders_products_merged.pkl"
path = r'/Users/elvachen/Documents/DA Learning/Immersion-data resources'
df_ods_prds_merged = pd.read_pickle(os.path.join(path, 'prepared data', 'orders_products_merged.pkl'))
df_ods_prds_merged.head(100)

# Create a subset dataframe with first one million rows
df_sub=df_ods_prds_merged[:1000000]

# Define a user-defined function to flag prices
def price_label(row):

  if row['prices'] <= 5:
    return 'Low-range product'
  elif (row['prices'] > 5) and (row['prices'] <= 15):
    return 'Mid-range product'
  elif row['prices'] > 15:
    return 'High range'
  else: return 'Not enough data'

# Create the new column by using the user-defined function
df_sub['price_range'] = df_sub.apply(price_label, axis=1)

# Check the outputs
df_sub['price_range'].value_counts(dropna=False)

# Check the maximum of the price
df_sub['prices'].max()

# If-statements with loc() function
df_sub.loc[df_sub['prices'] > 15, 'price_range_loc'] = 'High-range product'

df_sub.loc[(df_sub['prices'] <= 15) & (df_sub['prices'] > 5), 'price_range_loc'] = 'Mid-range product' 

df_sub.loc[df_sub['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df_sub['price_range_loc'].value_counts(dropna= False)

# Set price lable for entire dataframe by using if-statement  with loc()
df_ods_prds_merged.loc[df_ods_prds_merged['prices']>15, 'price_range_loc']= 'High-range product'

df_ods_prds_merged.loc[(df_ods_prds_merged['prices']<=15) & (df_ods_prds_merged['prices'] > 5), 'price_range_loc']= 'Mid-range product'

df_ods_prds_merged.loc[df_ods_prds_merged['prices'] <= 5, 'price_range_loc']='Low-range product'

df_ods_prds_merged['price_range_loc'].value_counts(dropna=False)

# for-loops example
for x in range(30, 45):
    print("My age is %d" % (x))

# Count the values of the column "orders_day_of_week"
df_ods_prds_merged['orders_day_of_week'].value_counts(dropna=False)

# Set the busy-tier label by using for-loops
result = []

for value in df_ods_prds_merged["orders_day_of_week"]:
  if value == 0:
    result.append("Busiest day")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

# Check the output
result

# Create a new column to contain the values in Result
df_ods_prds_merged['busiest_day'] = result

# Count the values in the new column
df_ods_prds_merged['busiest_day'].value_counts(dropna = False)

## Task of 4.7

# 4.7.2 Create a new column to label two busiest days and two slowest days
busy_tiers=[]

for value in df_ods_prds_merged['orders_day_of_week']:
    if value==0 :
        busy_tiers.append('First busiest day')
    elif value==1:
        busy_tiers.append('Second busiest day')
    elif value==4:
        busy_tiers.append('First lowest day')
    elif value==3:
        busy_tiers.append('Second lowest day')
    else:
        busy_tiers.append('Regular day')

df_ods_prds_merged['busiest_days']=busy_tiers
df_ods_prds_merged['busiest_days'].value_counts()

# 4.7.3 Check the values of this new column for accuracy. Note any observations in markdown format

According to the numbers of values in the new column, the result is correct. The numbers of "First buiest day", "Second buiest day", "First lowest day" and "Second lowest day" should respectively be the numbers of 4, 3, 0, 1 in column "orders_day_of_week". The number of "Regular day" should be the sum amount of 2, 5, 6 in column "orders_day_of_week".

# 4.7.4 Create a new column containing "Most orders", "Average orders", and "Fewest orders" labels called “busiest_period_of_day.”
    # Count the values of "order_hour_of_day"
df_ods_prds_merged['order_hour_of_day'].value_counts()

    # Create a variable to contain the result. 
    #In this scenario, consider hour 9-16 as "Most orders", hour 2-5 as "Fewest orders", besides, as "Average orders"
hour_tier=[]

for value in df_ods_prds_merged['order_hour_of_day']:
        if value in range(9,17):
            hour_tier.append('Most orders')
        elif (value>=2) & (value<=5):
            hour_tier.append('Fewest orders')
        else:
            hour_tier.append('Average orders')

    #Create a new column "busiest_period_of_day" to contain the resule
df_ods_prds_merged['busiest_period_of_day']=hour_tier

# 4.7.5 Print the frequency for this new column
df_ods_prds_merged['busiest_period_of_day'].value_counts()

# Check the dataframe
df_ods_prds_merged.head(100)

# Drop the unnecessary columns
df_ods_prds_merged=df_ods_prds_merged.drop(['_merge', 'Unnamed: 0_x', 'Unnamed: 0_y', 'Unnamed: 0.1'], axis=1)

# 4.7.7 Export your dataframe as a pickle file
df_ods_prds_merged.to_pickle(os.path.join(path, 'Prepared Data', 'orders_products_merged_47.pkl'))
