==> 5 Analyzing - Grouping Data & Aggregating Variables.ipynb <==
## Examples of the class

# Import libraries
import pandas as pd
import numpy as np
import os

# Import the latest pickle file 
path = r'/Users/elvachen/Documents/DA Learning/Immersion-data resources'
df_ods_prds = pd.read_pickle(os.path.join(path, 'prepared data', 'orders_products_merged_47.pkl'))
df_ods_prds.head()

# Create a subset dataframe with the first one million rows in the original dataframe
df=df_ods_prds[:1000000]

# Check the shape
df.shape

# See the first 30 rows of the dataframe
df.head(10)

# Group "product_name" together
df.groupby('product_name')

# Group "department_id" and calculate mean for "order_number"
df.groupby('department_id').agg({'order_sequency': ['mean']})

# Another way to calculate mean
df.groupby('department_id')['order_sequency'].mean()

# Performing Multiple Aggregations
df.groupby('department_id').agg({'order_sequency': ['mean', 'min', 'max']})

# Aggregate with transform function
df_ods_prds['max_order'] = df_ods_prds.groupby(['user_id'])['order_sequency'].transform(np.max)

# Check the output
df_ods_prds.head(100)

# Change the number of display list in Jupyter
pd.options.display.max_rows = None

# Deriving column with loc()
df_ods_prds.loc[df_ods_prds['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'

df_ods_prds.loc[(df_ods_prds['max_order'] <= 40) & (df_ods_prds['max_order'] > 10), 'loyalty_flag'] = 'Regular customer'

df_ods_prds.loc[df_ods_prds['max_order'] <= 10, 'loyalty_flag'] = 'New customer'


# Count the numbers of the value in the new column
df_ods_prds['loyalty_flag'].value_counts(dropna= False)

# List specific columns
df_ods_prds[['user_id', 'loyalty_flag', 'order_sequency']].head(60)

## Task of 4.8

# 4.8.2 Find the aggregated mean of the “order_number” column grouped by “department_id” for entire dataframe
df_ods_prds.groupby('department_id')['order_sequency'].mean()

# 4.8.3 How do the results for the entire dataframe differ from those of the subset?
Compare to the subset, the result of entire dataframe showed means of all department_id, rather than parts of department_id. Also, the values of means changed a little, as the observation range of entire dataframe was larger than the subset.

# 4.8.5 The spending habits of the three types of customers you identified
df_ods_prds.groupby('loyalty_flag')['prices'].describe()

According to above output, Loyal-customer group spent the lowest average price with fewer range, while New-customer group spent the highest average price with wider range.

# 4.8.6 Create a spending flag for each user based on the average price across all their orders using specific criteria
    # Create a new column to contain average price for each customer
df_ods_prds['avg_price']=df_ods_prds.groupby('user_id')['prices'].transform(np.mean)
df_ods_prds.head(30)

    # Create a new column to conctain spending flag
df_ods_prds.loc[df_ods_prds['avg_price'] < 10, 'spending_flag'] = 'Low spender'

df_ods_prds.loc[df_ods_prds['avg_price'] >= 10, 'spending_flag'] = 'High spender'

# Check the output
df_ods_prds['spending_flag'].value_counts(dropna= False)

# 4.8.7 Create an order frequency flag that marks the regularity of a user’s ordering behavior according to the median in the “days_since_prior_order” column
    # Create a new column to contain median of 'days_since_prior_order'
df_ods_prds['median_days_s_p_ods']=df_ods_prds.groupby('user_id')['days_since_prior_order'].transform(np.median)
df_ods_prds.head(30)

    # Create a new column to contain frequency flag
fqcy_flag=[]

    for value in df_ods_prds['median_days_s_p_ods']:
        if value > 20:
            fqcy_flag.append('Non-frequent customer')
        elif (value > 10) & (value <= 20):
            fqcy_flag.append('Regular customer')
        else:
            fqcy_flag.append('Frequent customer')
df_ods_prds['frequency_flag']= fqcy_flag
df_ods_prds.head()

# Check to output
df_ods_prds['frequency_flag'].value_counts(dropna= False)

# 4.8.8 Export your dataframe as a pickle file
df_ods_prds.to_pickle(os.path.join(path, 'Prepared Data', 'orders_products_48.pkl'))
